add latest changes

+++ src/main/java/com/grcp/demo/votingapp/vote/domain/PoolOptionVotingResult.java
@@ -2,5 +2,5 @@ package com.grcp.demo.votingapp.vote.domain;
 
 import com.grcp.demo.votingapp.pool.domain.PoolOptionId;
 
-public record PoolOptionVotingResult(PoolOptionId poolOptionId, String description, Integer totalVotes) {
+public record PoolOptionVotingResult(PoolOptionId poolOptionId, String description, long totalVotes) {
 }
