+++ src/main/java/com/grcp/demo/votingapp/vote/entrypoint/model/VotingResultResponseDto.java
@@ -3,6 +3,6 @@ package com.grcp.demo.votingapp.vote.entrypoint.model;
 public record VotingResultResponseDto(
         Long optionId,
         String description,
-        Integer votes,
+        long votes,
         Double percentage) {
 }
