+++ src/main/java/com/grcp/demo/votingapp/vote/gateway/db/view/VotingResultView.java
@@ -1,4 +1,4 @@
 package com.grcp.demo.votingapp.vote.gateway.db.view;
 
-public record VotingResultView(Long poolOptionId, String description, Integer totalVotes) {
+public record VotingResultView(long poolOptionId, String description, long totalVotes) {
 }
